Fix issue on StrategyIndexed when a resource is not available
For far when a resources was not avaiable, all the selection would go to
the next available resource, then the balance could be wrong because one
resource will be selected twice (its own number and the number of the
previous resource)

Imagine we have resources: 1, 2, 3, 4

Then resource 2 is not available.

Every time it is selected due to an index, the final selection will
be 3 (the next in the list).

For now once a resource is unavailable it will be remove from the list
of candidates, then the balance will be right between all the available
resources.